+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Fix subskill bonus values being NaN
 - The 'Helpless' change flag and the pinned condition incorrectly removed dexterity penalties
 - Saving Throw bonuses from racial HD weren't being added
+- Armor and encumbrance weren't slowing down actors
 
 ## 0.77.4 (Hotfix)